Convert private Traptor role to logstash only
The Traptor role is now part of the public ansible-symphony repo.  The
private traptor role has been converted to “traptor-logstash”, and only
handles the logstash configuration for getting the data from kafka to
ES.  This functionality is deprecated in favor of the #pipeline
approach, but until that is a reality we still need this kafka to ES
logstash ansible code.